Relocate some helpers
