+++ src/helpers/helpers.canvas.js
@@ -1,5 +1,7 @@
 'use strict';
 
+import {isArray} from './helpers.core';
+
 const PI = Math.PI;
 const RAD_PER_DEG = PI / 180;
 const DOUBLE_PI = PI * 2;
@@ -197,4 +199,64 @@ export function _bezierCurveTo(ctx, previous, target, flip) {
 		flip ? target.controlPointNextY : target.controlPointPreviousY,
 		target.x,
 		target.y);
+}
+
+export function fontString(pixelSize, fontStyle, fontFamily) {
+	return fontStyle + ' ' + pixelSize + 'px ' + fontFamily;
+}
+
+export function measureText(ctx, data, gc, longest, string) {
+	var textWidth = data[string];
+	if (!textWidth) {
+		textWidth = data[string] = ctx.measureText(string).width;
+		gc.push(string);
+	}
+	if (textWidth > longest) {
+		longest = textWidth;
+	}
+	return longest;
+}
+
+export function longestText(ctx, font, arrayOfThings, cache) {
+	cache = cache || {};
+	var data = cache.data = cache.data || {};
+	var gc = cache.garbageCollect = cache.garbageCollect || [];
+
+	if (cache.font !== font) {
+		data = cache.data = {};
+		gc = cache.garbageCollect = [];
+		cache.font = font;
+	}
+
+	ctx.font = font;
+	var longest = 0;
+	var ilen = arrayOfThings.length;
+	var i, j, jlen, thing, nestedThing;
+	for (i = 0; i < ilen; i++) {
+		thing = arrayOfThings[i];
+
+		// Undefined strings and arrays should not be measured
+		if (thing !== undefined && thing !== null && isArray(thing) !== true) {
+			longest = measureText(ctx, data, gc, longest, thing);
+		} else if (isArray(thing)) {
+			// if it is an array lets measure each element
+			// to do maybe simplify this function a bit so we can do this more recursively?
+			for (j = 0, jlen = thing.length; j < jlen; j++) {
+				nestedThing = thing[j];
+				// Undefined strings and arrays should not be measured
+				if (nestedThing !== undefined && nestedThing !== null && !isArray(nestedThing)) {
+					longest = measureText(ctx, data, gc, longest, nestedThing);
+				}
+			}
+		}
+	}
+
+	var gcLen = gc.length / 2;
+	if (gcLen > arrayOfThings.length) {
+		for (i = 0; i < gcLen; i++) {
+			delete data[gc[i]];
+		}
+		gc.splice(0, gcLen);
+	}
+	return longest;
 }
